simulator: fix pedantic C++ 98 warnings
C++ 98 requires all source files to terminate with a new-line character.
It also does not allow for semicolons outside of declarations that
require it (and in function scopes). Fix them to allow for pedantic
warnings to be enabled when beneficial. Finally, it does not allow for
object-to-function casts.

The first two warnings are resolved trivially. For the third, the
solution to cast through a pointer address is technically unsafe, as
systems are not required to support object pointers (e.g. void *)
being able to represent function pointers. However, the existence of
dlsym() virtually implies this is supported on the target system
(technically, all symbols could be data symbols, but that is
nitpicking).